Don't guess gateway and nameserver if not specified (closes: #28033)
Sometimes we don't want a default route to be set when configuring a
network interface with static address. The typical case is when we have
an external interface with dhcp and default route to the internet and a
static interface connected to the local network. See bug 28033 for
details.


SILENT: svn-revision: 465500